AnkiConnect API Call, Deck Data Structure
Started off by making sure I could call the
API from my device (not simulator). This in itself
proved to be a challenge because I was unaware I
needed to add an App Transport Security (ATS) setting
which allows local network access.
To load in decks, 2 separate API calls are used to first
fetch name and ids, then deck information respectively. I
created a DeckStructure node type which stores deck name, deck
information, and any other node children it may have. In the
overall data structure, there are 'n' amount of parent nodes
in the deckDataStructure array, each with their own children
and so on.
Network requests are made on the background thread with an UI
changes being put on @MainActor.