SwiftData Implementation, IP Address Persistance
First of all, changed project directory to better group
views, viewModels, and models.
In settingsView, IPAddress is now saved to user defaults and
is automatically used when syncing.
Changing DeckStructureNode to use SwiftData took quite a bit
of work... First of all, no longer need to use an array of
root nodes because we can query for root nodes using a
predicate and a new member of the DeckStructureNode class
'isRoot' to retrive all root nodes. When using list(children:),
I learned that for some reason SwiftData, when saving, initializes
optional relationships so that was leading to a problem with adding
more than 1 layer in the 'directory' of decks. To solve this, I
used a computer property that is basically a copy of the children
array, however if the array is empty, it returns nil, otherwise
a sorted array of children. Second, the list was still showing
the collapsable arrow even if the deck didn't have any subdecks.
I scowered the internet and found a post in the Apple Forums that I
needed to modify the relationships to have an inverse relationship.
So not only do the nodes have a children array, they also have a
relationship to it's parent. With both of these combined, I was able
to have a working directory of decks.